% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{Cg}
\alias{Cg}
\title{compute the potential gage capability (Cg)}
\usage{
Cg(sd_values, Tol, K = 0.2, num_sd = 6)
}
\arguments{
\item{sd_values}{the standard deviation of the MSA1 measurements}

\item{Tol}{the tolerance the gage shall be used for}

\item{K}{the correction factor for the tolerance (usually 0.2)}

\item{num_sd}{the numbers of standard deviations (default = 6)}
}
\value{
the potential gage Cg according to asmao script
}
\description{
compute the potential gage capability (Cg)
}
\examples{
Cg(1,50)
}
