% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{Cgk}
\alias{Cgk}
\title{Compute the actual gage capability with systematic error (Cgk)}
\usage{
Cgk(mean_values, sd_values, ref_value, Tol, K = 0.2, num_sd = 6)
}
\arguments{
\item{mean_values}{of the MSA1}

\item{sd_values}{of the MSA1}

\item{ref_value}{the value of the reference normal}

\item{Tol}{the tolerance in which the gage should measure}

\item{K}{the tolerance correction factor (default 0.2)}

\item{num_sd}{the numbers of standard deviations (default = 6)}
}
\value{
the Cgk value as a number
}
\description{
Compute the actual gage capability with systematic error (Cgk)
}
\examples{
Cgk(19,0.5,21,17)
}
