% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_LCL}
\alias{cmp_LCL}
\title{cmoputer lower control limit}
\usage{
cmp_LCL(dataset, n = 5)
}
\arguments{
\item{dataset}{data}

\item{n}{size of the subgroup}
}
\value{
the lower control limit
}
\description{
cmoputer lower control limit
}
