% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_LCL_s}
\alias{cmp_LCL_s}
\title{compute the lower control limit for s chart}
\usage{
cmp_LCL_s(dataset, n = 5)
}
\arguments{
\item{dataset}{data}

\item{n}{size of the subgroup}
}
\value{
the lower control limit
}
\description{
compute the lower control limit for s chart
}
