% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_LWL}
\alias{cmp_LWL}
\title{compute lower warning limit}
\usage{
cmp_LWL(dataset, n = 5)
}
\arguments{
\item{dataset}{dataset}

\item{n}{size of the subgroup}
}
\value{
the lower warning limit
}
\description{
compute lower warning limit
}
