% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_LWL_s}
\alias{cmp_LWL_s}
\title{compute the lower warning limit for s chart}
\usage{
cmp_LWL_s(dataset, n = 5)
}
\arguments{
\item{dataset}{data}

\item{n}{size of the subgroup}
}
\value{
the lower warning limit
}
\description{
compute the lower warning limit for s chart
}
