% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_ppm_cpk.R
\name{cmp_ppm_cpk}
\alias{cmp_ppm_cpk}
\title{Compute ppm from given Cpk}
\usage{
cmp_ppm_cpk(process_cpk)
}
\arguments{
\item{process_cpk}{the process Cpk as calculated by Cpk = (min(mu - LSL | USL - mu))/(6*sd) --> classical approach}
}
\value{
A value rounded to the nearest integer giving the ppm
}
\description{
Compute ppm from given Cpk
}
\examples{
ppm <- cmp_ppm_cpk(2)
}
