% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_s_mean}
\alias{cmp_s_mean}
\title{compute the mean of s (s chart)}
\usage{
cmp_s_mean(dataset, n = 5)
}
\arguments{
\item{dataset}{data}

\item{n}{size of the subgroup}
}
\value{
the s mean
}
\description{
compute the mean of s (s chart)
}
