% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_UCL}
\alias{cmp_UCL}
\title{compute upper control limit}
\usage{
cmp_UCL(dataset, n = 5)
}
\arguments{
\item{dataset}{the data}

\item{n}{size of the subgroup}
}
\value{
the upper control limit
}
\description{
compute upper control limit
}
