% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_UCL_s}
\alias{cmp_UCL_s}
\title{compute upper control limit for s chart}
\usage{
cmp_UCL_s(dataset, n = 5)
}
\arguments{
\item{dataset}{data}

\item{n}{subgroup}
}
\value{
the upper control limit
}
\description{
compute upper control limit for s chart
}
