% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_UCL_US}
\alias{cmp_UCL_US}
\title{Compute Upper Control Limit (US)}
\usage{
cmp_UCL_US(dataset, n = 5)
}
\arguments{
\item{dataset}{the dataset upon which to compute}

\item{n}{the size of the subgroup}
}
\value{
the UCL
}
\description{
Compute Upper Control Limit (US)
}
