% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_UWL}
\alias{cmp_UWL}
\title{Compute Upper Warning Limit}
\usage{
cmp_UWL(dataset, n = 5)
}
\arguments{
\item{dataset}{dataset}

\item{n}{size of the subgruop}
}
\value{
the upper warning limit
}
\description{
Compute Upper Warning Limit
}
