% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{cmp_UWL_s}
\alias{cmp_UWL_s}
\title{compute the upper warning limit for s chart}
\usage{
cmp_UWL_s(dataset, n = 5)
}
\arguments{
\item{dataset}{the data}

\item{n}{size of the subgroup}
}
\value{
the upper warning limit
}
\description{
compute the upper warning limit for s chart
}
