% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{comp_variants}
\alias{comp_variants}
\title{Compute possible variants in manufacturing}
\usage{
comp_variants(
  num_products = 5,
  num_tools_per_product = 5,
  total_num_available_tools = 7
)
}
\arguments{
\item{num_products}{number of products to manufacture}

\item{num_tools_per_product}{average number of tools used per product}

\item{total_num_available_tools}{total number of available tools}
}
\value{
a simulated dartaframe
}
\description{
Compute possible variants in manufacturing
}
\examples{
 dataset <- comp_variants()
}
