% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{Cp}
\alias{Cp}
\title{compute the potential process capability (Cp)}
\usage{
Cp(sd_values, Tol, num_sd = 6)
}
\arguments{
\item{sd_values}{the process standard deviation}

\item{Tol}{the tolerance window in which to compute}

\item{num_sd}{the numbers of standard deviations (default = 6)}
}
\value{
the Cp values according to asmao script
}
\description{
compute the potential process capability (Cp)
}
\examples{
Cp(1,10)
}
