% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{Cpk}
\alias{Cpk}
\title{compute the actual process capability (Cpk)}
\usage{
Cpk(mean_values, sd_values, usl, lsl, num_sd = 6)
}
\arguments{
\item{mean_values}{the mean value of the process}

\item{sd_values}{the standard deviation of the process}

\item{usl}{the upper specification limit}

\item{lsl}{the lower specification limit}

\item{num_sd}{the numbers of standard deviations (default = 6)}
}
\value{
the actual process capability Cpk according to asmao script
}
\description{
compute the actual process capability (Cpk)
}
\examples{
Cpk(19,0.5,21,17)
}
