% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_voetsch_cycle.R
\name{cut_voetsch_cycle}
\alias{cut_voetsch_cycle}
\title{Cut a SensorLab Voetsch dataset into its cycles}
\usage{
cut_voetsch_cycle(
  dataset,
  voetsch_total_number_of_cycles,
  voetsch_cycle = list(step_no = c(1, 2, 3, 4, 5, 6), temperature_celsius = c(-40, -40,
    150, 150, 23, 23), duration_min = c(0, 30, 30, 15, 30, 15))
)
}
\arguments{
\item{dataset}{a dataset as normalized by read_voetsch}

\item{voetsch_total_number_of_cycles}{what is the total number of cycles in the dataset? --> usually repetitions or someone canceled it}

\item{voetsch_cycle}{a list that specifies the programmed voetsch cycle. See default values for spec. Can probably refer to a standardized cycle at some point?}
}
\value{
a the original dataframe including the cycle number and a measurement index within that cycle
}
\description{
Cut a SensorLab Voetsch dataset into its cycles
}
\examples{
\dontrun{
out <- cut_voetsch_cycle(dataset, voetsch_total_number_of_cycles = 33)
}
}
