% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_sigend_rank.R
\name{dwilcox_signed_rank}
\alias{dwilcox_signed_rank}
\title{PMF calculation fo the sigend rank wilcox test for small sample sizes}
\usage{
dwilcox_signed_rank(n, lower_percentile = 0.025, upper_percentile = 0.975)
}
\arguments{
\item{n}{Sample Size}

\item{lower_percentile}{the lower percentile for the significance level (here 0.025 as default for 0.05/2}

\item{upper_percentile}{the upper percentile for the significance level (here 0.975 as default for 100-(0.05/2)}
}
\value{
a list of pmf with critical values
}
\description{
PMF calculation fo the sigend rank wilcox test for small sample sizes
}
\examples{
out = dwilcox_signed_rank(10)

}
