% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_ml.R
\name{ggconmat}
\alias{ggconmat}
\title{ggplot function to draw a confusion matrix based on the caret package}
\usage{
ggconmat(cm)
}
\arguments{
\item{cm}{a confusion matrix as returned by the caret package/function confusionMatrix}
}
\value{
a ggplot object
}
\description{
ggplot function to draw a confusion matrix based on the caret package
}
\examples{
\dontrun{
ggconmat(cm)
}

}
