% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plt.R
\name{plt_rule_of_ten}
\alias{plt_rule_of_ten}
\title{plot rule of ten: ggplot function to plot the rule of ten}
\usage{
plt_rule_of_ten()
}
\value{
a ggplot object
}
\description{
plot rule of ten: ggplot function to plot the rule of ten
}
\examples{
\dontrun{
plt_rule_of_ten()
}

}
