% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_voetsch.R
\name{read_voetsch}
\alias{read_voetsch}
\title{read raw data from voetsch twt in the sensor lab}
\usage{
read_voetsch(path_to_fn)
}
\arguments{
\item{path_to_fn}{path to the *.csv file from voetsch}
}
\value{
a dataframe with date_time conversion already done
}
\description{
read raw data from voetsch twt in the sensor lab
}
\examples{
\dontrun{
out <- read_voetsch("inst/extdata/USER-2024-04-04.csv")
}
}
