% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fun.R
\name{sig_fun}
\alias{sig_fun}
\title{sigmoid function}
\usage{
sig_fun(x, beta_0, beta_1)
}
\arguments{
\item{x}{independent var}

\item{beta_0}{parameter 1}

\item{beta_1}{parameter 2}
}
\value{
a value
}
\description{
A sigmoid function of the form f(x) = 1/(1-e^-(beta_0 + beta_1*x))
}
\examples{
y <- sig_fun(1,1,1)
}
